Complete support for child scopes. Locking rules for child scopes is as follows:
1. Rollback undoes all local changes and releases all acquired locks and lock upgrades. It behaves as if it never existed.
2. Commit promotes all acquired locks and lock upgrades to parent. It behaves as if it was never a child scope.

Releasing non-exclusive locks was considered, but this breaks expected behavior for repeatable-read isolation. A child scope might read a record, and then decide to insert a record based on it. Outside the scope, the parent might read the record again, which must be the same. Otherwise, it might choose to insert a conflicting record.