Pages can be deleted for two reasons: An existing page is dirtied, or a page truly isn't used any more. Any request to dirty a page can fail when the storage device is full, and the caller must abort the request. Deleting pages as part of a dirty request is essentially a soft request.
For a true page delete, it must succeed, or else the page can get lost. It's difficult for the caller to abort and cleanup because it likely went past a point of no return, or the page is being deleted to cleanup a failed attempt of something else.

True deletes now pass a new "force" option, and they just let the append heap grow indefinitely if necessary. A future checkpoint might be able to persist the pages, but it's also possible that no future checkpoints can complete. An OutOfMemoryError is possible with the force option, and so a failed forced delete attempt should always panic the database.

Recycling of pages is always treated as a forced operation, because recycling is never performed as part of a dirty request. It should always try to succeed, because the caller isn't expecting a recycle operation to fail. Any failed recycle request should also panic the database.