The exchange/insert/replace/update/deleteGhost operations must be atomic, even when using the BOGUS transaction. The internal find method supports a mode in which the leaf latch is retained, but is held shared. To complete the operation, the latch must be upgraded to exclusive. If this fails, the latch must be released and re-acquired. This permits the original value to change, and so atomicity would be lost.
The fix is to ditch capturing the original value during the find, and only capture it after the leaf is held exclusively. The original value could still be captured during the find, but if the latch upgrade fails then it would need to be double captured. The new code must not capture the original value from a split node. The find method can search into a split node, which is why the original code didn't need to call the notSplitDirty method.