Fix handling of prepared transactions when leadership is lost. A special lock is defined which is used to track prepared transactions. To safely transfer prepared transactions, the state must be copied into a new transaction, passed to the ReplEngine, and then the prepare lock is released. All of this is necessary because transactions cannot be safely shared among threads.
When the database becomes a replica, it waits for all prepared transactions to be transferred before accepting new changes. This is necessary if any changes are received which apply to prepared transactions. If the transaction isn't known by ReplEngine, it creates a new transaction, and then the state is broken.

A new event is defined which reports on stuck transactions, starting with a one second timeout, and then doubling up to one minute. Ideally the transaction's attachment should be reported too, but this isn't expected to be discoverable until after the transaction has been transferred. But at that point, it's not stuck anymore.