Revert change #ec3a36cb4ee5fb54b28b3717efd3c716ca408194  "Trigger now extends Clutch instead of CommitLock", and introduce a WideLatch class, which is a non-reentrant variant of CommitLock. It's a bit faster because it doesn't need to maintain thread-local state.
Shared acquisition with WideLatch or CommitLock is faster than Clutch, although Clutch is designed for adaptability if exclusive requests are common. It's expected that replacing the Trigger will be extremely rare, only when the set of secondary indexes changes. So the cost of acquiring exclusive access or constructing new instances isn't a concern.